fix(indexer): decay idle token/trader windows and flag stale overview (#577)

Closes #577

See merge request PlasticDigits/cl8y-dex-terraclassic!381

+++ indexer/src/api/tokens.rs
@@ -132,7 +132,7 @@ pub async fn get_token(
             volume: v.volume.to_string(),
             volume_usd: v.volume_usd.to_string(),
             trade_count: v.trade_count,
-            unique_traders: v.unique_traders,
+            unique_traders: i64::from(v.unique_traders),
         })
         .collect();
